make TimeSetting Class
