+++ app/src/main/java/com/example/dt/MainActivity_Juyeong.kt
@@ -64,26 +64,10 @@ class MainActivity_Juyeong : AppCompatActivity() {
 
         time.setOnClickListener {
             // 휴대폰 사용 시간 제어
-            // 시작 시간 선택
-            val startHour = sharedPreferences.getInt("startHour", 0)
-            val startMinute = sharedPreferences.getInt("startMinute", 0)
-            val timePickerDialogStart = TimePickerDialog(this, { _, hourOfDay, minute ->
-                sharedPreferences.edit().putInt("startHour", hourOfDay).putInt("startMinute", minute).apply()
-
-                // 종료 시간 선택
-                val endHour = sharedPreferences.getInt("endHour", 0)
-                val endMinute = sharedPreferences.getInt("endMinute", 0)
-                val timePickerDialogEnd = TimePickerDialog(this, { _, endHourOfDay, endMinute ->
-                    sharedPreferences.edit().putInt("endHour", endHourOfDay).putInt("endMinute", endMinute).apply()
-                }, endHour, endMinute, true)
-
-                timePickerDialogEnd.setTitle("종료 시간 선택")
-                timePickerDialogEnd.show()
-
-            }, startHour, startMinute, true)
-
-            timePickerDialogStart.setTitle("시작 시간 선택")
-            timePickerDialogStart.show()
+            val TimeSetting = TimeSetting(this)
+            time.setOnClickListener {
+                TimeSetting.showTimePickerDialog()
+            }
         }
 
         use_time.setOnClickListener {
